+++ app/allQuizzes/page.tsx
@@ -1,0 +1,10 @@
+
+export default function allQuizzes() {
+    return (
+        <div className="flex-1 w-full flex flex-col gap-20 items-center">
+            <div className="animate-in flex-1 flex flex-col gap-20 opacity-0 max-w-4xl px-3">
+                Quizzes
+            </div>     
+        </div>
+    )
+}
