+++ app/page.tsx
@@ -1,7 +1,7 @@
-import AuthButton from "../components/AuthButton";
 import { createClient } from "@/utils/supabase/server";
-import Header from "@/components/Header";
+import Nav from "@/components/Nav";
 import { redirect } from "next/navigation";
+import Footer from "@/components/Footer";
 
 export default async function Index() {
   const supabase = createClient();
@@ -18,22 +18,9 @@ export default async function Index() {
 
   return (
     <div className="flex-1 w-full flex flex-col gap-20 items-center">
-      <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
-        <div className="w-full max-w-4xl flex justify-between items-center p-3 text-sm">
-          <AuthButton />
-        </div>
-      </nav>
-
       <div className="animate-in flex-1 flex flex-col gap-20 opacity-0 max-w-4xl px-3">
-        <Header />
-      </div>
-
-      <footer className="w-full border-t border-t-foreground/10 p-8 flex justify-center text-center text-xs">
-        <p>
-          Powered by{" "}
-          <b>Algorhythm</b>
-        </p>
-      </footer>
+         Content
+      </div>     
     </div>
   )
 }
