--- components/Header.tsx
@@ -1,10 +1,0 @@
-
-export default function Header() {
-  return (
-      <header>
-        <div>
-          <h1>SoCBrain</h1>
-        </div>
-      </header>
-  );
-}
