+++ components/Nav.tsx
@@ -1,0 +1,37 @@
+import { createClient } from "@/utils/supabase/server";
+import Link from "next/link";
+import { redirect } from "next/navigation";
+
+export default async function Nav() {
+  const supabase = createClient();
+
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const signOut = async () => {
+    "use server";
+
+    const supabase = createClient();
+    await supabase.auth.signOut();
+    return redirect("/login");
+  };
+
+  return user ? (
+    <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
+        <div className="w-full max-w-4xl flex justify-between items-center p-3 text-sm">
+            <div className="flex items-center gap-4">
+              <Link href='/'>Home</Link>
+              <Link href='/allQuizzes'>Quizzes</Link>
+              <Link href='https://bot.schoolofcode.com/' target="_blank">SoCBot</Link>
+                Hey, {user.email}!
+                <form action={signOut}>
+                    <button className="py-2 px-4 rounded-md no-underline bg-btn-background hover:bg-btn-background-hover">
+                        Logout
+                    </button>
+                </form>
+            </div>
+        </div>
+    </nav>
+  ) : null
+}
